+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [BaseContext.alluka][yuyo.interactions.BaseContext.alluka] property for getting the
+  context's alluka client.
+- [pagination.ResponseKwargs][yuyo.pagination.ResponseKwargs] is now publicly exported to
+  allow 3rd party implementations of
+  [AbstractPage.ctx_to_kwargs][yuyo.pagination.AbstractPage.ctx_to_kwargs]
+  and [AbstractPage.to_kwargs][yuyo.pagination.AbstractPage.to_kwargs] to have properly
+  typed return types.
+
 ## [1.22.0] - 2024-07-11
 ### Added
 - Support for component pagination based on global ("static") bot data.